pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "backmax-mini-app"
version = "0.2.0"
description = "Skeleton services for the MAX mini app "
readme = "README.md"
authors = [{ name = "Blinchiki Team" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0,<0.112.0",
    "uvicorn[standard]>=0.30.0,<0.31.0",
    "pydantic>=2.8.0,<2.9.0",
    "pydantic-settings>=2.3.0,<2.4.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "sqlalchemy[asyncio]>=2.0.0,<3.0.0",
    "asyncpg>=0.29.0,<0.30.0",
    "fastapi-admin>=1.0.4,<1.1.0",
    "tortoise-orm>=0.25.1,<0.26.0",
    "redis>=5.0.0,<6.0.0",
    "python-multipart>=0.0.20,<0.1.0",
]
